secrets-manager/requirements.txt
certifi==2024.2.2
charset-normalizer==3.3.2
click==8.1.7
ibm-cloud-sdk-core==3.19.2
ibm-platform-services==0.51.1
ibm-secrets-manager-sdk==2.1.3
idna==3.6
markdown-it-py==3.0.0
mdurl==0.1.2
Pygments==2.17.2
PyJWT==2.8.0
python-dateutil==2.8.2
python-dotenv==1.0.1
requests==2.31.0
rich==13.7.1
six==1.16.0
urllib3==2.2.1
